backend/app/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-core
langchain-community
langchain-chroma
langchain-huggingface
langchain-groq
chromadb
pypdf
duckduckgo-search
wikipedia
arxiv
validators
youtube-transcript-api
unstructured
pymongo
motor